Use config on backup device - not a centralized config.
Previously, a single configuration had to cover every
backup situation imaginable. However, if the config
is read from the device itself, then each device can
specify its own settings. Thus an external HD
can backup everything, and a USB flash
can restrict itself to critical data.

Having a centralized config to set defaults still works.
However, one downside is that a CHANGE in the default
configuration can't trickle down to existing
devices. [Example: adding Paratext9]
The change will need to be made to each device.
This downside can be mitigated if we copy
backupDirs/backupInclude to ~/.config/wasta-backup
   -replaces the function of prevBackupDate.
   -allows easy verification/comparison of backup configs
   -allows easy config recovery in case device is lost
A separate commit will do this copying.

With the main config on the backup device, it seems
a little too dangerous/limiting to do a --delete rsync
when filling in missing files from the central config,
so that was removed.