remove menu item: Advanced - onlyBackupImportant
The only purpose for this was to allow a large
backup device to be able to backup everything,
even if the config files indicated that it should
only include specified extensions.

Since the backup device can now contain its own
configuration, there is no need for this
(IMHO) confusing option.

Added some tr() code to alert the user when it isn't
a full backup - a replacement for the banner
notice that needed to be removed.