replace prevBackupDate with copies of backup configs
The prevBackupDate.txt file had no real purpose.
The timestamp on the copied backup configs can
serve the same purpose. At the same time, we have
a copy of the config in case the device is lost.
We can also use something like MELD to compare
the different configs that are in use.

prevBackupDevice.txt does have a use in
selecting a USB-device when wasta-backup is
run manually. Otherwise I would have axed that
one as well.

Thanks to Nate Marti for this suggestion.