remove historical-config files before sync
Synchronizing from the default configs got more complicated
when we used the backup as the active-configuration source.
So we don't want the real configs to be updated (therefore
rsync --ignore-existing), but there IS value in getting
a copy of the lastUsedConfigs (for restoreAll as well as
easy meld-comparison), and certainly at least we don't
want stale "<first>UsedConfigs" (since they would never get updated).

So just delete these log-type files containing historical info.
Then the following sync will update with the central copy.

An alternative would be running two rsyncs, but that makes the log
messy.