explain how to make a valid wasta-backup location.
The normal use is with a USB drive, and in that case the program
auto-created the missing subfolder. However, if the user wants to
backup to a non-usb device, there isn't anything in the documentation
indicating what needs to happen. So document that in the error message.
That keeps the documentation clean/simple, and provides the hint
when it is most needed.