1.3.3 bugfix: enclose paths in double-quotes
Ahh, the standard bug that I failed to notice when I copied
rsync commands.  Of course, a space in the USB name
(or anywhere else, but that is the most likely although
the username might also have a space...) needs to
be handled with quoting.

Thanks to Steve Miller for reporting this.